requirements.txt
openai==0.28.0
python-telegram-bot==20.3
gspread>=5.7.2
oauth2client>=4.1.3
gspread-formatting>=1.1.4
pytz>=2023.3
apscheduler>=3.10.1
httpx>=0.24.0
python-dotenv>=1.0.0
colorlog>=6.6.0
